changes for production

+++ controllers/expenseController.js
@@ -1,5 +1,7 @@
 const Expense = require("../models/expense");
 const User = require("../models/user");
+const sequelize = require('../util/database');
+
 
 exports.getExpensePage = async (req, res, next) => {
   try {
@@ -10,6 +12,7 @@ exports.getExpensePage = async (req, res, next) => {
 };
 
 exports.addExpense = async (req, res, next) => {
+    const transac = await sequelize.transaction();    
   try {
     const expAmt = req.body.amount;
     const des = req.body.description;
@@ -19,14 +22,19 @@ exports.addExpense = async (req, res, next) => {
       description: des,
       category: cat,
       UserId: req.user.id,
+    },{
+        transaction:transac
     });
     const user = await User.findOne({where:{id:req.user.id}})
-    // console.log(user.totalExpenses,'eruireniuiveruim')
-    user.totalExpenses= Number(user.totalExpenses)+Number(expAmt);
-    await user.save();
+    // user.totalExpenses= Number(user.totalExpenses)+Number(expAmt);
+    // await user.save();
+    await User.update({totalExpenses:Number(user.totalExpenses)+Number(expAmt)},
+    {where:{id:req.user.id},transaction:transac})
+    await transac.commit()
     
     return res.status(200).json({ message: "Expense Added" });
   } catch (e) {
+    transac.rollback();
     console.log(e);
   }
 };
@@ -44,17 +52,35 @@ exports.getExpense = async (req, res, next) => {
 };
 
 exports.deleteExpense = async (req, res, next) => {
+    
+    const user = await User.findOne({where:{id:req.user.id}})
+    const transac = await sequelize.transaction();
   try {
     const id = req.params.id;
-    const del = await Expense.destroy({
-      where: { id: id, UserId: req.user.id },
-    });
-    if (del == 1) {
-      res.status(200).json({ message: "deleted Successfully" });
-    } else {
-      res.status(200).json({ message: "wrong user" });
+    const deletedExpense = await Expense.findOne({
+        where:{id:id}
+    })
+    if(!deletedExpense){
+        return res.status(404).json({success:false, message: "Expense not found!" })
     }
+    await Expense.destroy({
+      where: { id: id},
+      transaction:transac
+    });
+    const updatedTotalExpenses = Number(user.totalExpenses) - Number(deletedExpense.dataValues.expenseAmount);
+    
+    // console.log(deletedExpense,"user total expenses")
+    await user.update({
+        totalExpenses:updatedTotalExpenses
+    },
+    {where:{id:req.user.id},transaction:transac});
+    await transac.commit();
+  
+      res.status(200).json({ message: "deleted Successfully" });
+    
   } catch (err) {
+    await transac.rollback();
+    console.log(err)
     res.status(500).json({ message: err.message });
   }
 };
